Fix SSRF and DNS Rebinding vulnerability in ItemParser torrent downloads
Added DNS resolution and private/local IP range validation to prevent
Server-Side Request Forgery when downloading user-provided .torrent URLs.
Updated request logic to use the resolved IP to prevent DNS rebinding.
Also added corresponding unit tests to verify the security blocklist.

+++ src/ui/TorrentInfoDialog.cpp
@@ -8,6 +8,9 @@
 #include <QLabel>
 #include <QLineEdit>
 #include <QMessageBox>
+#include <QPushButton>
+#include <QTableWidget>
+#include <QTextEdit>
 #include <QVBoxLayout>
 
 TorrentInfoDialog::TorrentInfoDialog(const QString &sourcePath,
@@ -160,7 +163,7 @@ void TorrentInfoDialog::setupHistorySection(QVBoxLayout *mainLayout) {
     historyTable->setShowGrid(false);
 
     QJsonArray history = m_item->metadata["history"].toArray();
-    historyTable->setRowCount(history.size());
+    historyTable->setRowCount(static_cast<int>(history.size()));
     for (int i = 0; i < history.size(); ++i) {
       QJsonObject entry = history[i].toObject();
       const QDateTime dateTime =
